fix: type casting

+++ src/engine/render_system.cpp
@@ -42,7 +42,7 @@ namespace mitmeo
                     blit::Point(s.w / 2, s.h / 2),
                     s.scale,
                     s.transform);
-                // Origin
+                // Debug draw origin
                 blit::screen.line(blit::Point(p.x - 2, p.y), blit::Point(p.x + 2, p.y));
                 blit::screen.line(blit::Point(p.x, p.y - 2), blit::Point(p.x, p.y + 2));
             }
@@ -54,11 +54,12 @@ namespace mitmeo
                 auto &c = colliding.get<components::Collider>(e);
                 auto &p = colliding.get<components::Position>(e);
                 auto r = blit::Rect{
-                    p.x + c.x,
-                    p.y + c.y,
-                    c.w,
-                    c.h};
+                    p.x + (int32_t)c.x,
+                    p.y + (int32_t)c.y,
+                    (int32_t)c.w,
+                    (int32_t)c.h};
 
+                // Debug draw colliders
                 blit::screen.line(blit::Point(r.x - r.w / 2, r.y - r.h / 2), blit::Point(r.x + r.w / 2, r.y - r.h / 2));
                 blit::screen.line(blit::Point(r.x - r.w / 2, r.y + r.h / 2), blit::Point(r.x + r.w / 2, r.y + r.h / 2));
                 blit::screen.line(blit::Point(r.x - r.w / 2, r.y - r.h / 2), blit::Point(r.x - r.w / 2, r.y + r.h / 2));
